Track spline update
Added the dedicated track spline mesh as well as some qol updates to the spline to make it easier to see when editing it.

In addition the collision is changed to now work as intended with the mesh that is used.

Currently the spline fails to close properly with the mesh and when the spline is long enough, editing it while the mesh is attached will cause unreal to lag quite badly.

however these are things that shouldn't cause any real problems as the tracks are most likely not going to be entirely on a spline and the mesh can be easily detached and reattached when needing to edit the spline.