Adding cars, functional
added a base class that allows for copy pasting values into it at a bare minimum to function.

currently the variables don't update the actual vehicle for some reason but as it stands it works and can be used even if some of the base values need to be tweaked for feeling sake but this can be done at ease later on.

for now after this more functionality will be added now that it finally works.