AI Cars Now Working
IT WORKS, IT FUCKING WORKS!

Checkpoints are now in an array system attached to a spline so that checkpoints aren't put down one by one.
Checkpoints automatically detect what position they are and send the total to the the manager and cars.
Checkpoint array allows for selecting specific checkpoint styles for chosen checkpoints i.e gate / no gate.
Checkpoint array allows for target speed to be chosen for AI so that they can drive slower towards it if in a bad spot.
Checkpoints allow for one to be skipped at a time and for checkpoints to be toggled as a shortcut so AI will willingly skip the next checkpoint.

AI Now fully works and can be competitive if done right.
AI will drift if it detects its angle is far away from where it should be pointing
AI will adhere to target speed given by checkpoints
AI will reset when it has stopped moving or has flipped onto its side or back.

Theoretically compatible with infinite AI cars but my code is real damn messy so don't do more than like 10 or so.

While it is still fully functional, it is still unintuitive for our artists so holding back from pushing to main until it is.

HOLY FUCK THE CARS DRIVE HOLY FUCK, HOLY FUCK! :Pogchamp: